small renames for clarity
