+++ src/components/MainNav/BottomNav/BottomNav.js
@@ -2,28 +2,28 @@ import * as s from "./BottomNav.sc";
 import { useLocation } from "react-router-dom/cjs/react-router-dom";
 import { useEffect, useState, useRef } from "react";
 import NavIcon from "../NavIcon";
-import MoreOptions from "./MoreOptions";
-import BottomNav_Student from "./BottomNav_Student";
-import BottomNav_Teacher from "./BottomNav_Teacher";
+import MoreOptionsPanel from "./MoreOptionsPanel";
+import BottomNavOptionsForStudent from "./BottomNavOptionsForStudent";
+import BottomNavOptionsForTeacher from "./BottomNavOptionsForTeacher";
 import BottomNavOption from "./BottomNavOption";
 import strings from "../../../i18n/definitions";
 import { fadeIn, fadeOut } from "../../transitions";
 import { slideIn, slideOut } from "../../transitions";
 
+const PAGES_WITHOUT_BOTTOM_NAV = ["/exercises", "/read"];
 export default function BottomNav({ isOnStudentSide, isTeacher }) {
   const path = useLocation().pathname;
 
-  const [overlayTransition, setOverlayTransition] = useState();
+  const [overlayTransition, setOverlayTransition] = useState({});
   const [renderMoreOptions, setRenderMoreOptions] = useState(false);
-  const [moreOptionsTransition, setMoreOptionsTransition] = useState();
-  const [bottomNavTransition, setBottomNavTransition] = useState();
+  const [moreOptionsTransition, setMoreOptionsTransition] = useState({});
+  const [bottomNavTransition, setBottomNavTransition] = useState({});
   const [renderBottomNav, setRenderBottomNav] = useState(false);
 
-  const delay = useRef();
+  const delay = useRef(0); // initialize with an int that will represent the timeout id
 
   useEffect(() => {
-    clearTimeout(delay.current);
-    if (path.includes("/exercises") || path.includes("/read")) {
+    if (PAGES_WITHOUT_BOTTOM_NAV.includes(path)) {
       setBottomNavTransition(slideOut);
       delay.current = setTimeout(() => setRenderBottomNav(false), 500);
     } else {
@@ -57,12 +57,13 @@ export default function BottomNav({ isOnStudentSide, isTeacher }) {
             $isOnStudentSide={isOnStudentSide}
           >
             {isOnStudentSide && (
-              <BottomNav_Student isOnStudentSide={isOnStudentSide} />
+              <BottomNavOptionsForStudent isOnStudentSide={isOnStudentSide} />
             )}
 
             {!isOnStudentSide && (
-              <BottomNav_Teacher isOnStudentSide={isOnStudentSide} />
+              <BottomNavOptionsForTeacher isOnStudentSide={isOnStudentSide} />
             )}
+
             <BottomNavOption
               linkTo={"/account_settings"}
               currentPath={path}
@@ -82,7 +83,7 @@ export default function BottomNav({ isOnStudentSide, isTeacher }) {
       )}
 
       {renderMoreOptions && (
-        <MoreOptions
+        <MoreOptionsPanel
           currentPath={path}
           isOnStudentSide={isOnStudentSide}
           isTeacher={isTeacher}
