+++ src/components/MainNav/BottomNav/MoreOptionsPanel.sc.js
